--- src/main/java/com/bangbangbwa/backend/domain/member/common/Role.java
@@ -1,5 +1,0 @@
-package com.bangbangbwa.backend.domain.member.common;
-
-public enum Role {
-  MEMBER, BROADCASTER, ADMIN
-}
