+++ src/main/java/com/bangbangbwa/backend/domain/member/common/enums/Role.java
@@ -1,0 +1,5 @@
+package com.bangbangbwa.backend.domain.member.common.enums;
+
+public enum Role {
+  MEMBER, BROADCASTER, ADMIN
+}
