+++ src/main/java/com/bangbangbwa/backend/domain/oauth/business/OAuthFeignManager.java
@@ -1,9 +1,9 @@
 package com.bangbangbwa.backend.domain.oauth.business;
 
-import com.bangbangbwa.backend.domain.oauth.common.GoogleInfoDto;
-import com.bangbangbwa.backend.domain.oauth.common.KakaoInfoDto;
-import com.bangbangbwa.backend.domain.oauth.common.NaverInfoDto;
-import com.bangbangbwa.backend.domain.oauth.common.OAuthInfoDto;
+import com.bangbangbwa.backend.domain.oauth.common.dto.GoogleInfoDto;
+import com.bangbangbwa.backend.domain.oauth.common.dto.KakaoInfoDto;
+import com.bangbangbwa.backend.domain.oauth.common.dto.NaverInfoDto;
+import com.bangbangbwa.backend.domain.oauth.common.dto.OAuthInfoDto;
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
