+++ src/main/java/com/bangbangbwa/backend/domain/oauth/business/feign/NaverFeign.java
@@ -1,4 +1,4 @@
-package com.bangbangbwa.backend.domain.oauth.business;
+package com.bangbangbwa.backend.domain.oauth.business.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
