--- src/main/java/com/bangbangbwa/backend/domain/oauth/common/NaverInfoDto.java
@@ -1,5 +1,0 @@
-package com.bangbangbwa.backend.domain.oauth.common;
-
-public class NaverInfoDto {
-
-}
