--- src/main/java/com/bangbangbwa/backend/domain/oauth/common/SnsType.java
@@ -1,5 +1,0 @@
-package com.bangbangbwa.backend.domain.oauth.common;
-
-public enum SnsType {
-  KAKAO, GOOGLE, NAVER
-}
