+++ src/main/java/com/bangbangbwa/backend/domain/oauth/common/dto/GoogleInfoDto.java
@@ -1,4 +1,4 @@
-package com.bangbangbwa.backend.domain.oauth.common;
+package com.bangbangbwa.backend.domain.oauth.common.dto;
 
 public record GoogleInfoDto(
     String sub,
