+++ src/main/java/com/bangbangbwa/backend/domain/oauth/common/dto/KakaoInfoDto.java
@@ -1,0 +1,5 @@
+package com.bangbangbwa.backend.domain.oauth.common.dto;
+
+public class KakaoInfoDto {
+
+}
