+++ src/main/java/com/bangbangbwa/backend/domain/oauth/common/enums/SnsType.java
@@ -1,0 +1,5 @@
+package com.bangbangbwa.backend.domain.oauth.common.enums;
+
+public enum SnsType {
+  KAKAO, GOOGLE, NAVER
+}
