+++ src/main/java/com/bangbangbwa/backend/domain/oauth/service/OAuthService.java
@@ -1,8 +1,8 @@
 package com.bangbangbwa.backend.domain.oauth.service;
 
 import com.bangbangbwa.backend.domain.oauth.business.OAuthInfoProvider;
-import com.bangbangbwa.backend.domain.oauth.common.OAuthInfoDto;
-import com.bangbangbwa.backend.domain.oauth.common.SnsType;
+import com.bangbangbwa.backend.domain.oauth.common.dto.OAuthInfoDto;
+import com.bangbangbwa.backend.domain.oauth.common.enums.SnsType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
